[DDW-280] Separating files
