+++ storybook/stories/WalletScreens.stories.js
@@ -10,189 +10,23 @@ import { withKnobs, text, boolean, number, select } from '@storybook/addon-knobs
 import BigNumber from 'bignumber.js';
 import moment from 'moment';
 import startCase from 'lodash/startCase';
-import faker from 'faker';
 
 // Assets and helpers
-import Provider from './support/StoryProvider';
+import StoryLayout from './support/StoryLayout';
+import StoryProvider from './support/StoryProvider';
 import StoryDecorator from './support/StoryDecorator';
-import { generateRandomTransaction, generateTransaction, generateAddres } from './support/utils';
-import walletsIcon from '../../source/renderer/app/assets/images/sidebar/wallet-ic.inline.svg';
-import settingsIcon from '../../source/renderer/app/assets/images/sidebar/settings-ic.inline.svg';
-import adaIcon from '../../source/renderer/app/assets/images/sidebar/ada-redemption-ic.inline.svg';
-import paperCertificateIcon from '../../source/renderer/app/assets/images/sidebar/paper-certificate-ic.inline.svg';
+import { generateTransaction, generateAddres, promise } from './support/utils';
 import { formattedWalletAmount } from '../../source/renderer/app/utils/ada/formatters';
-import NodeSyncStatusIcon from '../../source/renderer/app/components/widgets/NodeSyncStatusIcon';
-import WalletAddress from '../../source/renderer/app/domains/WalletAddress';
 import { transactionStates, transactionTypes } from '../../source/renderer/app/domains/WalletTransaction';
-import Wallet from '../../source/renderer/app/domains/Wallet.js'
 import { assuranceModeOptions } from '../../source/renderer/app/types/transactionAssuranceTypes.js';
-
-// Empty screen elements
-import TopBar from '../../source/renderer/app/components/layout/TopBar';
-import Sidebar from '../../source/renderer/app/components/sidebar/Sidebar';
-import SidebarLayout from '../../source/renderer/app/components/layout/SidebarLayout';
-import SidebarWalletsMenu from '../../source/renderer/app/components/sidebar/wallets/SidebarWalletsMenu';
-import WalletWithNavigation from '../../source/renderer/app/components/wallet/layouts/WalletWithNavigation';
 
 // Screens
 import WalletSummary from '../../source/renderer/app/components/wallet/summary/WalletSummary';
 import WalletSendForm from '../../source/renderer/app/components/wallet/WalletSendForm';
 import WalletReceive from '../../source/renderer/app/components/wallet/WalletReceive';
 import WalletTransactionsList from '../../source/renderer/app/components/wallet/transactions/WalletTransactionsList';
-import WalletSettings from '../../source/renderer/app/components/wallet/WalletSettings';
+// import WalletSettings from '../../source/renderer/app/components/wallet/WalletSettings';
 
-const WALLETS = [
-  {
-    id: '0',
-    name: 'With Password',
-    amount: new BigNumber(0),
-    assurance: assuranceModeOptions.NORMAL,
-    hasPassword: true,
-    passwordUpdateDate: moment().subtract(1, 'month').toDate(),
-  },
-  {
-    id: '1',
-    name: 'No Password',
-    amount: new BigNumber(66.998),
-    assurance: assuranceModeOptions.NORMAL,
-    hasPassword: false,
-    passwordUpdateDate: new Date(),
-  }
-];
-
-const SIDEBAR_WALLETS = [
-  { id: WALLETS[0].id, title: WALLETS[0].name, info: `${WALLETS[0].amount} ADA`, isConnected: true },
-  { id: WALLETS[1].id, title: WALLETS[1].name, info: `${WALLETS[1].amount} ADA`, isConnected: true },
-];
-
-const sidebarCategories = [
-  {
-    name: 'WALLETS',
-    route: '/wallets',
-    icon: walletsIcon,
-  },
-  // {
-  //   name: 'SETTINGS',
-  //   route: '/settings',
-  //   icon: settingsIcon,
-  // },
-  {
-    name: 'ADA_REDEMPTION',
-    route: '/ada-redemption',
-    icon: adaIcon,
-  },
-  {
-    name: 'PAPER_WALLET',
-    route: '/paper-wallet-create-certificate',
-    icon: paperCertificateIcon,
-  }
-];
-
-const sidebarMenus = observable({
-  wallets: {
-    items: SIDEBAR_WALLETS,
-    activeWalletId: '1',
-    actions: {
-      onAddWallet: action('toggleAddWallet'),
-      onWalletItemClick: (walletId: string) => {
-        runInAction(() => sidebarMenus.wallets.activeWalletId = walletId);
-      }
-    }
-  }
-});
-
-const validateAmount = (amountInNaturalUnits: string) => (
-  new Promise(resolve => {
-    setTimeout(() => {
-      resolve(amountInNaturalUnits ? true : false);
-    }, 2000);
-  })
-);
-const calculateTransactionFee = (receiver: string, amount: string) => (
-  new Promise(resolve => {
-    setTimeout(() => {
-      resolve((receiver || amount) ? true : false);
-    }, 2000);
-  })
-);
-
-type Props = {
-  storyName?: string,
-  children?: any | Node
-};
-
-@observer
-class WalletScreen extends Component<Props> {
-
-  render() {
-
-    const { children, storyName='' } = this.props;
-    const activeNavItem = storyName.split(' ')[0].toLowerCase();
-
-    return (
-      <div
-        style={{
-          height: '100vh'
-        }}
-      >
-        <SidebarLayout
-          sidebar={this.getSidebar(!!children)}
-          topbar={this.getTopbar(activeNavItem)}
-        >
-          {
-            storyName !== 'Empty' &&
-              (
-                <WalletWithNavigation
-                  isActiveScreen={item => item === activeNavItem}
-                  onWalletNavItemClick={linkTo('WalletScreens', item => startCase(item))}
-                >
-                  {children}
-                </WalletWithNavigation>
-              )
-          }
-        </SidebarLayout>
-      </div>
-    );
-  }
-
-  @observable
-  isShowingSubMenus = !!this.props.children;
-
-  getSidebar = () => (
-    <Sidebar
-      categories={sidebarCategories}
-      activeSidebarCategory={sidebarCategories[0].route}
-      menus={sidebarMenus}
-      isShowingSubMenus={this.isShowingSubMenus}
-      onCategoryClicked={action('onCategoryClicked')}
-      isDialogOpen={() => false}
-      onAddWallet={action('onAddWallet')}
-      openDialogAction={action('openDialog')}
-      onSubmitSupportRequest={()=>{}}
-    />
-  );
-
-  getTopbar = (activeNavItem: string) => (
-    <TopBar
-      onToggleSidebar={() => {
-        runInAction(() => this.isShowingSubMenus = !this.isShowingSubMenus);
-      }}
-      formattedWalletAmount={formattedWalletAmount}
-      currentRoute={`/wallets/${WALLETS[sidebarMenus.wallets.activeWalletId].id}/${activeNavItem}`}
-      activeWallet={activeNavItem !== 'empty' ? new Wallet(WALLETS[sidebarMenus.wallets.activeWalletId]) : null}
-      showSubMenuToggle={true}
-      showSubMenus={this.isShowingSubMenus}
-    >
-      <NodeSyncStatusIcon
-        networkStatus={{
-          isSynced: true,
-          syncPercentage: 100,
-        }}
-        isProduction
-      />
-    </TopBar>
-  );
-}
 
 storiesOf('WalletScreens', module)
 
@@ -202,15 +36,15 @@ storiesOf('WalletScreens', module)
 
     return (
       <StoryDecorator>
-        <Provider
-          activeWallet={WALLETS[sidebarMenus.wallets.activeWalletId]}
+        <StoryProvider
+          activeWallet={{}}
         >
-          <WalletScreen
+          <StoryLayout
             storyName={context.story}
           >
             {storyWithKnobs}
-          </WalletScreen>
-        </Provider>
+          </StoryLayout>
+        </StoryProvider>
       </StoryDecorator>
     )
   })
@@ -242,8 +76,8 @@ storiesOf('WalletScreens', module)
       currencyUnit="Ada"
       currencyMaxFractionalDigits={ 6}
       currencyMaxIntegerDigits={11}
-      validateAmount={validateAmount}
-      calculateTransactionFee={calculateTransactionFee}
+      validateAmount={promise(true)}
+      calculateTransactionFee={promise(true)}
       addressValidator={()=>{}}
       openDialogAction={()=>{}}
       isDialogOpen={()=> boolean('hasDialog', false)}
@@ -262,7 +96,7 @@ storiesOf('WalletScreens', module)
       onGenerateAddress={()=>{}}
       onCopyAddress={()=>{}}
       isSidebarExpanded
-      walletHasPassword={WALLETS[sidebarMenus.wallets.activeWalletId].hasPassword}
+      walletHasPassword={boolean('walletHasPassword', false)}
       isSubmitting={boolean('isSubmitting', false)}
     />
   ))
